bmw_driveiq/requirements.txt
groq>=0.9.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
streamlit>=1.35.0
plotly>=5.18.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
joblib>=1.3.0
scipy>=1.11.0
